Requested changes (Moved endpoint from workspace to configuration)
